move custom input fields to separate file
